fix(stream): XREADGROUP history re-read bumps PEL delivery count + refreshes time — p5t3n
fr's Store::xreadgroup history-read path (explicit start ID, not '>') left the
PEL untouched, with a comment claiming "Redis does NOT increment delivery count".
That's wrong: upstream t_stream.c streamReplyWithRangeFromConsumerPEL does
`nack->delivery_count++` and `nack->delivery_time = now` for every served PEL
entry whose underlying stream entry STILL EXISTS. Tombstoned entries (emitted as
[id, nil], handled by s0614 — carried here as an empty field list) take the
[id, NULL] branch and are NOT touched.

So after a consumer re-read its history (XREADGROUP ... 0), fr's XPENDING
delivery counts and idle times drifted below redis. Now each live re-served PEL
entry bumps deliveries and refreshes last_delivered_ms; tombstones are skipped.

Found via scripts/stream_xinfo_differ.py — a new seeded randomized stream-
sequence differ that applies identical XADD/XDEL/XTRIM/XSETID/XGROUP/XREADGROUP/
XACK sequences to fr and the vendored redis 7.2.4 oracle and compares normalized
XINFO STREAM FULL + XINFO GROUPS after every op. Verified vs redis: live re-read
delivery 1->2->3, tombstone'd entry stays 1, idle resets on re-read. Updated two
tests (fr-store xpending_summary, fr-command xpending detail) that had encoded
the old no-bump behavior.

frankenredis-p5t3n

+++ crates/fr-command/src/lib.rs
@@ -36553,18 +36553,22 @@ mod tests {
         assert_eq!(
             detail,
             RespFrame::Array(Some(vec![
-                // Pending replay does NOT increment delivery count or update idle time
+                // The c1 '0' history replay at t=20 re-served the live PEL
+                // entries 1000-0/1000-1, so — matching upstream
+                // streamReplyWithRangeFromConsumerPEL — their delivery counts
+                // bumped to 2 and delivery times refreshed to 20 (idle now
+                // 40-20=20). 1000-2 was delivered once to c2 at t=30 (idle 10).
                 RespFrame::Array(Some(vec![
                     RespFrame::BulkString(Some(b"1000-0".to_vec())),
                     RespFrame::BulkString(Some(b"c1".to_vec())),
-                    RespFrame::Integer(30), // idle since initial delivery at t=10
-                    RespFrame::Integer(1),  // delivery count stays at 1
+                    RespFrame::Integer(20), // idle since the t=20 replay refresh
+                    RespFrame::Integer(2),  // delivered once + one history re-read
                 ])),
                 RespFrame::Array(Some(vec![
                     RespFrame::BulkString(Some(b"1000-1".to_vec())),
                     RespFrame::BulkString(Some(b"c1".to_vec())),
-                    RespFrame::Integer(30),
-                    RespFrame::Integer(1),
+                    RespFrame::Integer(20),
+                    RespFrame::Integer(2),
                 ])),
                 RespFrame::Array(Some(vec![
                     RespFrame::BulkString(Some(b"1000-2".to_vec())),
@@ -36595,14 +36599,14 @@ mod tests {
                 RespFrame::Array(Some(vec![
                     RespFrame::BulkString(Some(b"1000-0".to_vec())),
                     RespFrame::BulkString(Some(b"c1".to_vec())),
-                    RespFrame::Integer(30),
-                    RespFrame::Integer(1),
+                    RespFrame::Integer(20),
+                    RespFrame::Integer(2),
                 ])),
                 RespFrame::Array(Some(vec![
                     RespFrame::BulkString(Some(b"1000-1".to_vec())),
                     RespFrame::BulkString(Some(b"c1".to_vec())),
-                    RespFrame::Integer(30),
-                    RespFrame::Integer(1),
+                    RespFrame::Integer(20),
+                    RespFrame::Integer(2),
                 ])),
             ]))
         );
